patch #4342

+++ src/utils/lib/tiger.c
@@ -779,7 +779,7 @@ char hexa(int i)
 #define MAX_TIGER_CHUNK_SIZE 1024
 static word64 tiger_buffer[MAX_TIGER_CHUNK_SIZE];
 
-void tiger_hash(char prefix, char *s, int len, unsigned char *digest)
+void tiger_hash(char prefix, char *s, long len, unsigned char *digest)
 {
   char *buffer = (char*) tiger_buffer;
   word64 ndigest[3];
@@ -794,10 +794,10 @@ void tiger_hash(char prefix, char *s, int len, unsigned char *digest)
   swap_digest(digest);
 }
 
-void tiger_tree_string(char *s, int len, int pos, int block_size, char *digest)
+void tiger_tree_string(char *s, size_t len, size_t pos, size_t block_size, char *digest)
 {
   if(block_size == BLOCK_SIZE){
-    int length = (len - pos > BLOCK_SIZE) ? BLOCK_SIZE : len - pos;
+    size_t length = (len - pos > BLOCK_SIZE) ? BLOCK_SIZE : len - pos;
     tiger_hash(0, s+pos, length, digest);
   } else {    
     if(pos+block_size/2 >=len){
@@ -827,9 +827,9 @@ void tiger_tree_string(char *s, int len, int pos, int block_size, char *digest)
 #include "caml/fail.h"
 #include "caml/alloc.h"
 
-int tiger_block_size(int len)
+long tiger_block_size(long len)
 {
-  int block_size = BLOCK_SIZE;
+  long block_size = BLOCK_SIZE;
   while(block_size < len) block_size *= 2;
   return block_size;
 }
